requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
shopify-python-api
facebook-business
python-multipart
ShopifyAPI
